pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "support-triage-env"
version = "1.0.0"
description = "Customer Support Ticket Resolution — an OpenEnv environment for LLM agent training"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }

dependencies = [
    "openenv-core>=0.2.3",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

# No scripts section — server is launched via Dockerfile CMD
# To run locally: uvicorn app:app --host 0.0.0.0 --port 8000 --reload
